+++ src/routes/portal/employee.route.ts
@@ -1,41 +1,58 @@
 import express from "express";
-import {
-	validateLeaveRequest,
-	createLeaveRequest,
-	viewLeaveRequest,
-} from "../../functions/portal/employee.function";
+import {makeLeaveRequest} from "../../functions/portal/post.portal.function";
+import {getLeaveRequest} from "../../functions/portal/get.portal.function";
+import {updateLeaveRequestStatus} from "../../functions/portal/patch.portal.function";
 const portal = express.Router();
 
-portal.post("/create/leave-request", async (req, res) => {
+/* 
+	Post requests
+*/
+portal.post("/make/leave-request", async (req, res) => {
+	const {body} = req;
 	try {
-		const company = req.employeeAuthCreds.employee;
-		const result = await createLeaveRequest(req.body, company);
+		const employee = req.employeeAuthCreds.employee;
+		const result = await makeLeaveRequest(body, employee);
 		res.status(result.status).json(result.data);
 	} catch (error: any) {
 		res.status(500).json({error: error.message});
 	}
 });
+/* 
+	Post requests
+*/
 
+/* 
+	Get requests
+*/
 portal.get("/view/leave-request/:leaveRequestUuid", async (req, res) => {
+	const {leaveRequestUuid} = req.params;
 	try {
-		const {leaveRequestUuid} = req.params;
-		const result = await viewLeaveRequest(leaveRequestUuid);
+		const result = await getLeaveRequest(leaveRequestUuid);
 		res.status(result.status).json(result.data);
 	} catch (error: any) {
 		res.status(500).json({error: error.message});
 	}
 });
+/* 
+	Get requests
+*/
 
-portal.patch("/validate/leave-request/:leaveRequestUuid", async (req, res) => {
+/* 
+	Patch requests
+*/
+portal.patch("/update/respond/leave-request/:leaveRequestUuid", async (req, res) => {
+	const {leaveRequestUuid} = req.params;
+	const employee = req.employeeAuthCreds.employee;
+	const {body} = req;
 	try {
-		const {leaveRequestUuid} = req.params;
-		const company = req.employeeAuthCreds.employee;
-
-		const result = await validateLeaveRequest(leaveRequestUuid, company, req.body);
+		const result = await updateLeaveRequestStatus(leaveRequestUuid, employee, body);
 		res.status(result.status).json(result.data);
 	} catch (error: any) {
 		res.status(500).json({error: error.message});
 	}
 });
+/* 
+	Patch requests
+*/
 
 export default portal;
